Scratchpad: Added palette drawing feature
Provides a menu option in the Scratchpad tab to Draw a palette in the current
scratchpad. This will open a file dialog and let the user select a Gimp Palette
from the filesystem.

The colours within this palette will be used to draw a grid of colours, using
the current brush.